+++ packages/orq-rc/src/orq_ai_sdk/models/createeval_responsebody.py
@@ -277,9 +277,9 @@ class CreateEvalResponseBodyFunction(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-05-18T05:04:24.675Z"
+    created: Optional[str] = "2026-05-18T14:04:16.993Z"
 
-    updated: Optional[str] = "2026-05-18T05:04:24.675Z"
+    updated: Optional[str] = "2026-05-18T14:04:16.993Z"
 
     guardrail_config: OptionalNullable[
         CreateEvalResponseBodyEvalsResponseGuardrailConfig
@@ -500,9 +500,9 @@ class ResponseBodyPython(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-05-18T05:04:24.675Z"
+    created: Optional[str] = "2026-05-18T14:04:16.993Z"
 
-    updated: Optional[str] = "2026-05-18T05:04:24.675Z"
+    updated: Optional[str] = "2026-05-18T14:04:16.993Z"
 
     guardrail_config: OptionalNullable[CreateEvalResponseBodyEvalsGuardrailConfig] = (
         UNSET
@@ -725,9 +725,9 @@ class ResponseBodyHTTP(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-05-18T05:04:24.675Z"
+    created: Optional[str] = "2026-05-18T14:04:16.993Z"
 
-    updated: Optional[str] = "2026-05-18T05:04:24.675Z"
+    updated: Optional[str] = "2026-05-18T14:04:16.993Z"
 
     guardrail_config: OptionalNullable[CreateEvalResponseBodyGuardrailConfig] = UNSET
 
@@ -925,9 +925,9 @@ class ResponseBodyJSON(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-05-18T05:04:24.675Z"
+    created: Optional[str] = "2026-05-18T14:04:16.993Z"
 
-    updated: Optional[str] = "2026-05-18T05:04:24.675Z"
+    updated: Optional[str] = "2026-05-18T14:04:16.993Z"
 
     guardrail_config: OptionalNullable[ResponseBodyGuardrailConfig] = UNSET
 
@@ -1348,6 +1348,7 @@ class CreateEvalLLM2TypedDict(TypedDict):
     repetitions: NotRequired[int]
     categories: NotRequired[List[str]]
     categorical_labels: NotRequired[List[CreateEvalLLMEvalsCategoricalLabelsTypedDict]]
+    dataset_id: NotRequired[str]
 
 
 class CreateEvalLLM2(BaseModel):
@@ -1365,9 +1366,9 @@ class CreateEvalLLM2(BaseModel):
 
     jury: CreateEvalLLMJury
 
-    created: Optional[str] = "2026-05-18T05:04:24.675Z"
+    created: Optional[str] = "2026-05-18T14:04:16.993Z"
 
-    updated: Optional[str] = "2026-05-18T05:04:24.675Z"
+    updated: Optional[str] = "2026-05-18T14:04:16.993Z"
 
     guardrail_config: OptionalNullable[CreateEvalLLMEvalsResponseGuardrailConfig] = (
         UNSET
@@ -1379,6 +1380,8 @@ class CreateEvalLLM2(BaseModel):
 
     categorical_labels: Optional[List[CreateEvalLLMEvalsCategoricalLabels]] = None
 
+    dataset_id: Optional[str] = None
+
     @model_serializer(mode="wrap")
     def serialize_model(self, handler):
         optional_fields = set(
@@ -1389,6 +1392,7 @@ class CreateEvalLLM2(BaseModel):
                 "repetitions",
                 "categories",
                 "categorical_labels",
+                "dataset_id",
             ]
         )
         nullable_fields = set(["guardrail_config"])
@@ -1635,6 +1639,7 @@ class CreateEvalLLM1TypedDict(TypedDict):
     categorical_labels: NotRequired[
         List[CreateEvalLLMEvalsResponseCategoricalLabelsTypedDict]
     ]
+    dataset_id: NotRequired[str]
 
 
 class CreateEvalLLM1(BaseModel):
@@ -1652,9 +1657,9 @@ class CreateEvalLLM1(BaseModel):
 
     model: str
 
-    created: Optional[str] = "2026-05-18T05:04:24.675Z"
+    created: Optional[str] = "2026-05-18T14:04:16.993Z"
 
-    updated: Optional[str] = "2026-05-18T05:04:24.675Z"
+    updated: Optional[str] = "2026-05-18T14:04:16.993Z"
 
     guardrail_config: OptionalNullable[CreateEvalLLMEvalsGuardrailConfig] = UNSET
 
@@ -1666,6 +1671,8 @@ class CreateEvalLLM1(BaseModel):
         None
     )
 
+    dataset_id: Optional[str] = None
+
     @model_serializer(mode="wrap")
     def serialize_model(self, handler):
         optional_fields = set(
@@ -1676,6 +1683,7 @@ class CreateEvalLLM1(BaseModel):
                 "repetitions",
                 "categories",
                 "categorical_labels",
+                "dataset_id",
             ]
         )
         nullable_fields = set(["guardrail_config"])