+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -113,7 +113,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-05-18T05:04:23.600Z")
+    updated: Optional[datetime] = parse_datetime("2026-05-18T14:04:15.855Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")