+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KRWQHHWFH3QAMWPY9P0ZWZM7"
+        "01KRXPE5C8EY68C2J3QEJ7HETV"
     )
     r"""The unique identifier of the data source"""